# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Node.js 입문 주차 개인 과제 API",
    "description": "Node.js와 Express로 로그인 기능 없는 백엔드 서버 만들기",
    "version": "1.0.0"
  },
  "host": "http://3.34.196.250",
  "basePath": "",
  "schemes": [
    "http"
  ],
  "paths": {
    "/posts": {
      "post": {
        "tags": [
          "posts"
        ],
        "description": "게시글 작성",
        "parameters": [
          {
            "name": "",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "string",
                  "example": "Developer"
                },
                "password": {
                  "type": "string",
                  "example": "1234"
                },
                "title": {
                  "type": "string",
                  "example": "안녕하세요"
                },
                "content": {
                  "type": "string",
                  "example": "안녕하세요 content 입니다"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "게시글 작성 성공",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "게시글을 생성하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "body 또는 params를 입력받지 못한 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터 형식이 올바르지 않습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "posts"
        ],
        "description": "게시글 조회",
        "parameters": [],
        "responses": {
          "200": {
            "description": "게시글 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "postId": {
                        "type": "string",
                        "example": "639a81d11099c5ef762d3688"
                      },
                      "user": {
                        "type": "string",
                        "example": "Developer"
                      },
                      "title": {
                        "type": "string",
                        "example": "안녕하세요"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2022-12-14T03:39:20.389Z"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/posts/{_postId}": {
      "get": {
        "tags": [
          "posts"
        ],
        "description": "게시글 상세 조회",
        "parameters": [
          {
            "name": "_postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 상세 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "postId": {
                      "type": "string",
                      "example": "639a81d11099c5ef762d3688"
                    },
                    "user": {
                      "type": "string",
                      "example": "Developer"
                    },
                    "title": {
                      "type": "string",
                      "example": "안녕하세요"
                    },
                    "content": {
                      "type": "string",
                      "example": "안녕하세요 content 입니다."
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-12-14T03:39:20.389Z"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "body 또는 params를 입력받지 못한 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터 형식이 올바르지 않습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "posts"
        ],
        "description": "게시글 수정",
        "parameters": [
          {
            "name": "_postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "1234"
                },
                "title": {
                  "type": "string",
                  "example": "안녕하세요"
                },
                "content": {
                  "type": "string",
                  "example": "안녕하세요 content 입니다"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 수정 성공",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "게시글을 수정하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "body 또는 params를 입력받지 못한 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터 형식이 올바르지 않습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "비밀번호가 일치하지 않는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "게시글 수정 권한이 없습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "_postId에 해당하는 게시글이 존재하지 않을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "게시글 조회에 실패하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "description": "게시글 삭제",
        "parameters": [
          {
            "name": "_postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "1234"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 삭제 성공",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "게시글을 삭제하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "body 또는 params를 입력받지 못한 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터 형식이 올바르지 않습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "비밀번호가 일치하지 않는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "게시글 삭제 권한이 없습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "_postId에 해당하는 게시글이 존재하지 않을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "게시글 조회에 실패하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/comments/{_postId}": {
      "post": {
        "tags": [
          "comments"
        ],
        "description": "댓글 생성",
        "parameters": [
          {
            "name": "_postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "string",
                  "example": "Developer"
                },
                "password": {
                  "type": "string",
                  "example": "1234"
                },
                "content": {
                  "type": "string",
                  "example": "안녕하세요 댓글 입니다"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "댓글 작성 성공",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "댓글을 생성하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "body 또는 params를 입력받지 못한 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터 형식이 올바르지 않습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "찾을 수 없음"
          }
        }
      },
      "get": {
        "tags": [
          "comments"
        ],
        "description": "댓글 목록 조회",
        "parameters": [
          {
            "name": "_postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "댓글 목록 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "commentId": {
                        "type": "string",
                        "example": "639a82521099c5ef762d3698"
                      },
                      "user": {
                        "type": "string",
                        "example": "Developer"
                      },
                      "content": {
                        "type": "string",
                        "example": "안녕하세요 댓글입니다."
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2022-12-14T03:39:20.389Z"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "body 또는 params를 입력받지 못한 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터 형식이 올바르지 않습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/comments/{_commentId}": {
      "put": {
        "tags": [
          "comments"
        ],
        "description": "댓글 수정",
        "parameters": [
          {
            "name": "_commentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "1234"
                },
                "content": {
                  "type": "string",
                  "example": "수정된 댓글입니다."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "댓글 수정 성공",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "댓글을 수정하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "body 또는 params를 입력받지 못한 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터 형식이 올바르지 않습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "비밀번호가 일치하지 않는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "댓글 수정 권한이 없습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "_commentId에 해당하는 댓글이 존재하지 않을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "댓글 조회에 실패하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "comments"
        ],
        "description": "댓글 삭제",
        "parameters": [
          {
            "name": "_commentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "1234"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "댓글 삭제 성공",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "댓글을 삭제하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "body 또는 params를 입력받지 못한 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터 형식이 올바르지 않습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "비밀번호가 일치하지 않는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "댓글 삭제 권한이 없습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "_commentId에 해당하는 댓글이 존재하지 않을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "댓글 조회에 실패하였습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}